Create nothread
This script runs a web-scraper to find information on job listings from a fake job site, including the a link to further information on each job. It then web-scrapes each of these URLs to display a job description for each job. At the end, the processing time is printed out. (Uncomment line 45 to see the data frame generated by the script)